List the stored secrets in `cf service` output
We want users to feel confident updating the secrets stored by this
broker. A big step forward is for users to be able to see the names of
the secrets in each service instance.

This commit 'threads the needle' of how to make help text appear when
a user runs `cf service SERVICE_INSTANCE_NAME`.

In an ideal world I'd be able to configure the broker with
`instances_retrievable: true` and then Cloud Controller would get
helpful information back from the broker when someone ran
`cf service`, or at least after an operation. Unfortunately CC doesn't
do that yet [1].

Instead I have configured `Provision` and `Update` to claim to be
async operations. They are still completely synchronous, but claiming
to be async triggers Cloud Controller to then call the `LastOperation`
endpoint. `LastOperation` can return a human-readable message, and
that message is still displayed to end-users even once the operation
completes.

[1] https://github.com/cloudfoundry/cloud_controller_ng/issues/1504